Refactor main into namespaces in musicpp.h/ musicpp.cpp; Initial ui defines
