+++ include/amscraper.h
