+++ include/globals.h
@@ -1,0 +1,10 @@
+//
+// Created by Jonathan on 16-Jan-26.
+//
+
+#ifndef MUSICPP_GLOBALS_H
+#define MUSICPP_GLOBALS_H
+
+inline extern unsigned int g_wm_taskbar_restart = 0;
+
+#endif //MUSICPP_GLOBALS_H
