+++ src/musicpp.cpp
@@ -1,0 +1,865 @@
+//
+// Created by Jonathan on 16-Jan-26.
+//
+
+#include "musicpp.h"
+#include "timeutils.h"
+#include "stringutils.h"
+#include "globals.h"
+#include <chrono>
+#include <iostream>
+#include <spdlog/spdlog.h>
+#include <spdlog/sinks/basic_file_sink.h>
+#include <leveldb/db.h>
+#include <leveldb/write_batch.h>
+#include <consoleutils.h>
+#include <credhelper.h>
+#include <conio.h>
+#include <shlobj_core.h>
+
+using namespace std;
+
+string setup::MakeLogName() {
+    const auto t = UnixSecondsNow();
+
+    const chrono::zoned_time local{chrono::current_zone(), chrono::sys_time(chrono::seconds{t})};
+
+    const auto lt = chrono::floor<chrono::seconds>(local.get_local_time());
+    const auto dp = chrono::floor<chrono::days>(lt);
+    const chrono::year_month_day ymd{dp};
+    const auto tod = chrono::hh_mm_ss{lt - dp};
+
+    return std::format("musicpp_{:04}-{:02}-{:02}_{:02}-{:02}-{:02}.log",
+                       static_cast<int>(ymd.year()),
+                       static_cast<unsigned>(ymd.month()),
+                       static_cast<unsigned>(ymd.day()),
+                       tod.hours().count(),
+                       tod.minutes().count(),
+                       tod.seconds().count()
+    );
+}
+
+void setup::PruneLogs(const filesystem::path &folder) {
+    std::vector<filesystem::directory_entry> files;
+    for (auto &entry: filesystem::directory_iterator(folder)) {
+        if (entry.is_regular_file()) files.push_back(entry);
+    }
+    ranges::sort(files, [](auto &a, auto &b) {
+        return a.last_write_time() > b.last_write_time();
+    });
+
+    if (files.size() > kMaxFiles) {
+        for (size_t i = kMaxFiles; i < files.size(); i++) {
+            error_code ec;
+            filesystem::remove(files[i], ec);
+            if (ec) {
+                wcout << console::Yellow << L"Failed to remove old log file: " << ConvertToWString(
+                    files[i].path().string()) << console::Reset << endl;
+            }
+        }
+    }
+}
+
+bool setup::IsValidRegion(const std::string &region) {
+    return kValidRegions.contains(region);
+}
+
+void setup::PrintRegionList() {
+    wcout << L"\nAvailable Apple Music regions (country codes):" << endl;
+    int count = 0;
+    for (const auto &region: kRegionList) {
+        wcout << setw(4) << region.c_str();
+        ++count;
+        if (count % 12 == 0) {
+            wcout << endl;
+        }
+    }
+    if (count % 12 != 0) {
+        wcout << endl;
+    }
+    wcout << endl;
+}
+
+void setup::SaveRegion(const AppContext &ctx, const std::string &region) {
+    if (!ctx.db) return;
+    if (const leveldb::Status status = ctx.db->Put(leveldb::WriteOptions(), kRegionDbKey, region); !status.ok()) {
+        if (ctx.logger) ctx.logger->warn("Failed to persist region {}: {}", region, status.ToString());
+    } else {
+        if (ctx.logger) ctx.logger->info("Set region to {}", region);
+    }
+}
+
+string setup::LoadRegion(const AppContext &ctx) {
+    if (!ctx.db) return "";
+    string region;
+    if (const leveldb::Status status = ctx.db->Get(leveldb::ReadOptions(), kRegionDbKey, &region); !status.ok()) {
+        if (ctx.logger) ctx.logger->warn("Failed to read saved region: {}", status.ToString());
+    }
+    return region;
+}
+
+string setup::GetRegion() {
+    wstring input;
+    while (true) {
+        wcout << L"Enter Apple Music region code (type 'list' to view options): ";
+        wcin >> input;
+
+        string region_input = ToLowerCase(ConvertWString(input));
+        if (region_input == "list") {
+            PrintRegionList();
+            continue;
+        }
+
+        if (IsValidRegion(region_input)) {
+            return region_input;
+        }
+
+        wcout << L"Region not recognized. Please choose a code from the list below." << endl;
+        PrintRegionList();
+    }
+}
+
+string setup::EnsureRegion(const AppContext &ctx) {
+    string region = ToLowerCase(LoadRegion(ctx));
+    if (region.empty()) {
+        region = kDefaultRegion;
+        SaveRegion(ctx, region);
+    }
+    return region;
+}
+
+void setup::PurgeDatabase(const AppContext *ctx) {
+    if (!ctx->db) {
+        if (ctx->logger) ctx->logger->warn("Attempted to purge unavailable database");
+        return;
+    }
+
+    leveldb::WriteBatch batch;
+    leveldb::ReadOptions ro;
+    ro.fill_cache = false;
+    unique_ptr<leveldb::Iterator> it(ctx->db->NewIterator(ro));
+    if (!it) {
+        if (ctx->logger) ctx->logger->warn("Failed to get iterator for db purge");
+        return;
+    }
+
+    int count = 0;
+    for (it->SeekToFirst(); it->Valid(); it->Next()) {
+        const leveldb::Slice key = it->key();
+
+        if (key == kDiscordStateKey || key == kLfmStateKey || key == kRegionDbKey) {
+            continue;
+        }
+
+        batch.Delete(key);
+        count++;
+    }
+
+    if (!it->status().ok()) {
+        if (ctx->logger) ctx->logger->error("Iterator error during purge: {}", it->status().ToString());
+        return;
+    }
+
+    const leveldb::Status s = ctx->db->Write(leveldb::WriteOptions(), &batch);
+    if (s.ok()) {
+        if (ctx->logger) ctx->logger->info("Successfully wiped {} items from the cache", count);
+        if (count > 0) ctx->db->CompactRange(nullptr, nullptr);
+    } else {
+        if (ctx->logger) ctx->logger->error("Failed to delete items from cache! Error: {}", s.ToString());
+    }
+    ctx->player->ImageRefresh();
+}
+
+void setup::CleanDatabase(const AppContext &ctx) {
+    if (ctx.logger)
+        ctx.logger->info(
+            "Beginning database purge: checking for keys older than {}s (static) and keys older than {}s (animated)",
+            kDbExpireTime, kDbExpireTimeAnim);
+    wcout << L"Cleaning database of static images older than " << FriendlyTime(kDbExpireTime) << endl;
+    wcout << L"Cleaning database of animated images older than " << FriendlyTime(kDbExpireTimeAnim) << endl;
+    const uint64_t now = UnixSecondsNow();
+
+    leveldb::ReadOptions ro;
+    leveldb::WriteOptions wo;
+    ro.fill_cache = false;
+    wo.sync = true;
+
+    unique_ptr<leveldb::Iterator> it(ctx.db->NewIterator(ro));
+    if (!it) {
+        if (ctx.logger) ctx.logger->warn("Failed to get iterator for db purge");
+        wcout << console::Yellow << L"Error during db cleanup, check logs" << console::Reset << endl;
+        return;
+    }
+
+    leveldb::WriteBatch batch;
+
+    size_t deleted = 0;
+    size_t malformed = 0;
+
+    auto ParseEpochPrefix = [&](const string &value, uint64_t &out_seconds) {
+        const size_t sep = value.find('|');
+        if (sep == string::npos || sep == 0) return false;
+
+        out_seconds = std::stoull(value.substr(0, sep));
+        return true;
+    };
+
+    for (it->SeekToFirst(); it->Valid(); it->Next()) {
+        const leveldb::Slice key = it->key();
+        if (key == kDiscordStateKey || key == kLfmStateKey || key == kRegionDbKey) {
+            continue;
+        }
+
+        const leveldb::Slice val = it->value();
+        bool should_delete = false;
+
+        try {
+            uint64_t stored;
+            if (!ParseEpochPrefix(val.ToString(), stored)) {
+                should_delete = true;
+                malformed++;
+            } else {
+                if (stored > now) {
+                    should_delete = true;
+                    malformed++;
+                } else if (key.ToString().find("musicppAMAnim") != string::npos) {
+                    if (now - stored > kDbExpireTimeAnim) {
+                        should_delete = true;
+                    }
+                } else if (now - stored > kDbExpireTime) {
+                    should_delete = true;
+                }
+            }
+        } catch (exception &e) {
+            should_delete = true;
+            if (ctx.logger)
+                ctx.logger->warn("Found extremely malformed value (error: {}): {}", e.what(),
+                                 val.ToString());
+        }
+
+        if (should_delete) {
+            batch.Delete(key);
+            deleted++;
+        }
+    }
+
+    if (!it->status().ok()) {
+        if (ctx.logger) ctx.logger->error("Iterator error during database cleaning: {}", it->status().ToString());
+        wcout << console::Yellow << L"Error during db cleanup, check logs" << console::Reset << endl;
+        return;
+    }
+
+    if (deleted > 0) {
+        const leveldb::Status s = ctx.db->Write(wo, &batch);
+        if (!s.ok()) {
+            if (ctx.logger) ctx.logger->error("Failed to delete items from cache! Error: {}", s.ToString());
+            wcout << console::Yellow << L"Error during db cleanup, check logs" << console::Reset << endl;
+            return;
+        }
+        if (ctx.logger) ctx.logger->info("Purged {} keys ({} malformed)", deleted, malformed);
+        wcout << L"Purged " << deleted << L" keys from the database (" << malformed << L" malformed keys)" << endl;
+        ctx.db->CompactRange(nullptr, nullptr);
+    } else {
+        if (ctx.logger) ctx.logger->info("Nothing to purge");
+        wcout << L"Nothing to purge!" << endl;
+    }
+
+    wcout << endl;
+}
+
+bool setup::InitializeDatabase(AppContext &ctx) {
+    PWSTR path = nullptr;
+
+    if (FAILED(SHGetKnownFolderPath(FOLDERID_LocalAppData, 0, nullptr, &path))) {
+        wcout << console::Red << L"Fatal: could not get local appdata path" << console::Reset << endl;
+
+        return false;
+    }
+
+    wcout << L"Configuration files: " << path << L'\n' << endl;
+
+    const filesystem::path base_path(path);
+    const filesystem::path db_path = base_path / "musicpp" / "song_db";
+    const filesystem::path log_path = base_path / "musicpp" / "logs";
+    CoTaskMemFree(path);
+
+    std::error_code ec1, ec2;
+
+    create_directories(db_path, ec1);
+    create_directories(log_path, ec2);
+
+    if (ec1) {
+        wcout << console::Red << L"Fatal: Could not create db folder: " << ec1.message().c_str() << console::Reset
+                << endl;
+        return false;
+    }
+
+    if (ec2) {
+        wcout << console::Red << L"Could not create log folder: " << ec2.message().c_str() << console::Reset << endl;
+        wcout << L"The program will not have logs. Continuing..." << endl;
+    } else {
+        PruneLogs(log_path);
+        const string name = setup::MakeLogName();
+        const auto logger = spdlog::basic_logger_mt("MusicPP Logger", (log_path / name).string());
+#ifdef LOG_LEVEL_DEBUG
+        spdlog::set_level(spdlog::level::debug);
+        logger->flush_on(spdlog::level::debug);
+#elif defined(LOG_LEVEL_INFO)
+        spdlog::set_level(spdlog::level::info);
+        spdlog::flush_every(kLogFlushInterval);
+#else
+        spdlog::set_level(spdlog::level::debug);
+        logger->flush_on(spdlog::level::debug);
+#endif
+        spdlog::set_pattern("(%Y-%m-%d %H:%M:%S %z, [%8l]) Thread %t: %v");
+        // ReSharper disable once CppRedundantQualifierADL
+        spdlog::set_default_logger(logger);
+        ctx.logger = logger;
+    }
+
+    if (ctx.logger) {
+        ctx.logger->info("Logger initialized at path: {}", log_path.string());
+        ctx.logger->info("MusicPP starting up...");
+    }
+
+    leveldb::DB *temp_db = nullptr;
+
+    leveldb::Options options;
+
+    options.create_if_missing = true;
+
+    const leveldb::Status status = leveldb::DB::Open(options, db_path.string(), &temp_db);
+
+    if (status.ok()) {
+        ctx.db.reset(temp_db);
+        if (ctx.logger) ctx.logger->info("Database initialized at path: {}", db_path.string());
+        CleanDatabase(ctx);
+        return true;
+    }
+
+    wcout << console::Red << L"Fatal: Could not initialize database!" << console::Reset << endl;
+    if (ctx.logger) ctx.logger->error("Could not initialize database: {}", status.ToString());
+
+    return false;
+}
+
+bool setup::RunConfigurationMode(const AppContext &ctx, string &region) {
+    wcout << L"Press ANY KEY to enter Configuration/Reset mode." << endl;
+
+
+    bool interaction = false;
+
+    for (int i = 3; i > 0; --i) {
+        wcout << L"Launching in " << i << L" seconds... \r" << flush;
+
+        for (int j = 0; j < 20; ++j) {
+            if (_kbhit()) {
+                (void) _getch(); // consume key
+
+                interaction = true;
+
+                break;
+            }
+
+            this_thread::sleep_for(chrono::milliseconds(50));
+        }
+
+        if (interaction) break;
+    }
+
+
+    if (interaction) {
+        if (ctx.logger) ctx.logger->info("User interaction detected, entering Configuration Mode.");
+        wcout << console::Green << L"\n\n[Configuration Mode]" << console::Reset << endl;
+
+        wcout << L"Enter 1 to force reset all API Keys" << endl;
+
+        wcout << L"Enter 2 to change Apple Music region" << endl;
+
+        wcout << L"Enter anything else to exit" << endl;
+
+
+        wstring input;
+
+        wcin >> input;
+
+        if (input == L"1") {
+            if (ctx.logger) ctx.logger->info("User initiated API key reset.");
+            wcout << L"Deleted all keys." << endl;
+
+            return true;
+        }
+        if (input == L"2") {
+            wcout << L"Current Apple Music region: " << region.c_str() << endl;
+            PrintRegionList();
+
+            if (const string new_region = GetRegion(); new_region != region) {
+                region = new_region;
+                SaveRegion(ctx, region);
+                wcout << L"Apple Music region updated to " << region.c_str() << endl;
+            } else {
+                wcout << L"Region unchanged" << endl;
+            }
+        }
+    } else {
+        if (ctx.logger) ctx.logger->info("Auto-Start mode, no user interaction");
+        wcout << L"\n[Auto-Start] No interaction detected. Loading keys..." << endl;
+    }
+
+    return false;
+}
+
+void setup::LoadCredentials(AppContext &ctx, const bool &force_reset, const string &region) {
+    wcout << L"\nChecking API Keys..." << endl;
+
+    if (force_reset) {
+        DeleteGenericCredential(kLastFmDbSessionKey, ctx.logger.get());
+        DeleteGenericCredential(kSpotifyDbClientToken, ctx.logger.get());
+    }
+
+    const wstring s_cid = EnsureCredential(kSpotifyDbClientIdKey, L"Spotify Client ID",
+                                           L"https://developer.spotify.com/documentation/web-api/tutorials/getting-started",
+                                           force_reset, ctx.logger.get());
+
+    const wstring s_sec = EnsureCredential(kSpotifyDbClientSecretKey, L"Spotify Client Secret",
+                                           L"https://developer.spotify.com/documentation/web-api/tutorials/getting-started",
+                                           force_reset, ctx.logger.get());
+
+    const wstring i_cid = EnsureCredential(kImgurDbClientIdKey, L"Imgur Client ID",
+                                           L"https://api.imgur.com/oauth2/addclient", force_reset,
+                                           ctx.logger.get());
+
+    const wstring lfm_key = EnsureCredential(kLastFmDbApikey, L"LastFM API Key",
+                                             L"https://www.last.fm/api/account/create",
+                                             force_reset, ctx.logger.get());
+
+    const wstring lfm_secret = EnsureCredential(kLastFmDbSecret, L"LastFM Secret",
+                                                L"https://www.last.fm/api/account/create", force_reset,
+                                                ctx.logger.get());
+
+    wcout << L"All APIKeys found" << endl;
+    if (ctx.logger) {
+        ctx.logger->info("All API keys loaded successfully");
+    }
+
+    // Initialize API and Player objects
+    wcout << L"\nApple Music region: " << region.c_str() << endl;
+    ctx.scraper = std::make_unique<Amscraper>(region, ctx.logger.get());
+    if (ctx.logger) ctx.logger->info("AMScraper initialized with region {}", region);
+
+    ctx.processor = std::make_unique<M3U8Processor>(ctx.logger.get());
+    if (ctx.logger) ctx.logger->info("M3U8Processor initialized");
+
+    ctx.lastfm = std::make_unique<Lfm>(ConvertWString(lfm_key), ConvertWString(lfm_secret), ctx.db.get(),
+                                       ctx.logger.get());
+    if (ctx.logger) {
+        ctx.logger->info("LastFM initialized");
+    }
+
+    ctx.spotify = std::make_unique<SpotifyApi>(ConvertWString(s_cid), ConvertWString(s_sec), ctx.logger.get());
+    if (ctx.logger) {
+        ctx.logger->info("SpotifyAPI initialized");
+    }
+
+    ctx.imgur = std::make_unique<ImgurApi>(ConvertWString(i_cid), ctx.logger.get());
+    if (ctx.logger) ctx.logger->info("ImgurAPI initialized");
+
+    ctx.player = std::make_unique<MediaPlayer>(ctx.scraper.get(), ctx.processor.get(), ctx.spotify.get(),
+                                               ctx.imgur.get(),
+                                               ctx.lastfm.get(), ctx.db.get(), ctx.logger.get());
+    if (ctx.logger) ctx.logger->info("mediaPlayer initialized");
+
+    ctx.discord = std::make_unique<Discordrp>(ctx.player.get(), kDiscordApikey, ctx.db.get(), ctx.logger.get());
+    if (ctx.logger) ctx.logger->info("discordrp initialized");
+}
+
+variant<int, AppContext> setup::setup() {
+    SetupConsole();
+
+    AppContext ctx;
+
+    if (!InitializeDatabase(ctx)) {
+        // Let user see error before closing
+        std::wcout << L"Press Enter to continue..." << std::endl;
+        std::wcin.get();
+        return 1;
+    }
+
+    std::string region = EnsureRegion(ctx);
+    const bool force_reset = RunConfigurationMode(ctx, region);
+
+    LoadCredentials(ctx, force_reset, region);
+    if (ctx.logger) ctx.logger->info("Load credentials finished");
+
+    wcout << "\nThis console window will close automatically" << endl;
+
+    for (int i = 3; i > 0; --i) {
+        wcout << L"Closing in " << i << L" seconds... \r" << flush;
+
+        this_thread::sleep_for(chrono::seconds(1));
+    }
+
+    CleanupConsole();
+    if (ctx.logger) ctx.logger->info("Console cleanup complete, moving to message loop");
+    return ctx;
+}
+
+void loop::UpdateTrayTooltip(AppContext *ctx) {
+    if (ctx->nid.cbSize == 0) return;
+
+    const Snapshot metadata = ctx->player->GetSnapshot(kSnapshotTypeTime);
+
+    const wstring image_source = metadata.image_source;
+
+    const bool playing = metadata.state;
+
+    wstringstream tip_stream;
+
+    if (metadata.album.empty()) {
+        tip_stream << L"MusicPP\nNo track playing";
+    } else {
+        tip_stream << L"MusicPP\n";
+        if (ctx->discord->GetState()) {
+            tip_stream << L"Presence enabled\n";
+        }
+        if (ctx->lastfm->GetState()) {
+            tip_stream << L"LastFM enabled\n";
+        }
+        tip_stream << L"\nImage Source: " << (!image_source.empty() ? image_source : L"unknown") << L"\n";
+
+        if (playing) {
+            if (const uint64_t duration = metadata.duration; duration > 0) {
+                const uint64_t elapsed = metadata.elapsed;
+
+                tip_stream << L"" << FormatTimestamp(elapsed) << L" / " << FormatTimestamp(duration) << L"\n";
+            }
+        } else {
+            tip_stream << L"Paused" << L"\n";
+        }
+    }
+
+    const wstring tip = tip_stream.str();
+
+    if (tip == ctx->last_tip) {
+        return;
+    }
+
+    ctx->last_tip = tip;
+
+    if (ctx->logger) ctx->logger->debug("Updating tray tooltip to: \n{}", ConvertWString(tip));
+
+    lstrcpynW(ctx->nid.szTip, tip.c_str(), ARRAYSIZE(ctx->nid.szTip));
+
+    ctx->nid.uFlags |= NIF_TIP | NIF_ICON | NIF_MESSAGE;
+
+    Shell_NotifyIcon(NIM_MODIFY, &ctx->nid);
+}
+
+void loop::CopyToClipboard(const AppContext *ctx, const std::wstring &text, const wstring &label) {
+    if (!OpenClipboard(nullptr)) {
+        if (ctx->logger) ctx->logger->error("Failed to open clipboard for text: {}", ConvertWString(text));
+        return;
+    }
+    EmptyClipboard();
+
+    const size_t bytes = (text.size() + 1) * sizeof(wchar_t);
+    const HGLOBAL hMem = GlobalAlloc(GMEM_MOVEABLE, bytes);
+    if (!hMem) {
+        if (ctx->logger)
+            ctx->logger->error("Failed to allocate memory for clipboard text: {}",
+                               ConvertWString(text));
+        CloseClipboard();
+        return;
+    }
+
+    memcpy(GlobalLock(hMem), text.c_str(), bytes);
+    GlobalUnlock(hMem);
+    SetClipboardData(CF_UNICODETEXT, hMem);
+    CloseClipboard();
+
+    NOTIFYICONDATAW tempNid = {sizeof(tempNid)};
+    tempNid.hWnd = ctx->hWnd;
+    tempNid.uID = 1; // This must match the ID used in WM_CREATE (ctx->nid.uID)
+    tempNid.uFlags = NIF_INFO;
+    tempNid.dwInfoFlags = NIIF_INFO;
+    lstrcpynW(tempNid.szInfoTitle, L"Copied to clipboard", ARRAYSIZE(tempNid.szInfoTitle));
+    lstrcpynW(tempNid.szInfo, label.c_str(), ARRAYSIZE(tempNid.szInfo));
+
+    Shell_NotifyIcon(NIM_MODIFY, &tempNid);
+    if (ctx->logger) ctx->logger->debug("Set clipboard to: {}", ConvertWString(text));
+}
+
+LRESULT CALLBACK loop::WndProc(const HWND hWnd, const UINT msg, const WPARAM wParam, const LPARAM lParam) {
+    auto *ctx = reinterpret_cast<AppContext *>(GetWindowLongPtr(hWnd, GWLP_USERDATA));
+
+    if (msg == g_wm_taskbar_restart) {
+        if (ctx && ctx->nid.cbSize != 0) {
+            // Explorer restarted
+            ctx->nid.uFlags |= NIF_MESSAGE | NIF_ICON | NIF_TIP;
+            if (!ctx->nid.hIcon) {
+                ctx->nid.hIcon = LoadIcon(GetModuleHandle(nullptr), MAKEINTRESOURCE(IDI_APPICON));
+                if (!ctx->nid.hIcon) ctx->nid.hIcon = LoadIcon(nullptr, IDI_APPLICATION); // Fallback
+            }
+            Shell_NotifyIcon(NIM_ADD, &ctx->nid);
+            if (ctx->logger) ctx->logger->info("Shell restarted, tray icon recreated");
+        }
+
+        UpdateTrayTooltip(ctx);
+        return 0;
+    }
+
+    switch (msg) {
+        case WM_CREATE: {
+            const auto *create_struct = reinterpret_cast<CREATESTRUCT *>(lParam);
+
+            ctx = static_cast<AppContext *>(create_struct->lpCreateParams);
+
+            SetWindowLongPtr(hWnd, GWLP_USERDATA, reinterpret_cast<LONG_PTR>(ctx));
+
+            ctx->hWnd = hWnd;
+
+            ctx->nid.cbSize = sizeof(ctx->nid);
+
+            ctx->nid.hWnd = hWnd;
+
+            ctx->nid.uID = 1;
+
+            ctx->nid.uFlags = NIF_MESSAGE | NIF_ICON | NIF_TIP;
+
+            ctx->nid.uCallbackMessage = WM_TRAYICON;
+
+            ctx->nid.hIcon = LoadIcon(GetModuleHandle(nullptr), MAKEINTRESOURCE(IDI_APPICON));
+
+            if (!ctx->nid.hIcon) ctx->nid.hIcon = LoadIcon(nullptr, IDI_APPLICATION); // Fallback
+
+            lstrcpynW(ctx->nid.szTip, L"MusicPP", ARRAYSIZE(ctx->nid.szTip));
+
+            Shell_NotifyIcon(NIM_ADD, &ctx->nid);
+
+            if (ctx->logger) ctx->logger->info("Window created and tray icon added");
+
+            SetTimer(hWnd, DATA_TIMER_ID, kLoopRefreshInterval, nullptr);
+
+            if (ctx->logger)
+                ctx->logger->debug("Timer started: Data ({} ms)",
+                                   kLoopRefreshInterval);
+            return 0;
+        }
+
+
+        case WM_TRAYICON: {
+            if (LOWORD(lParam) == WM_RBUTTONUP) {
+                POINT pt{};
+
+                GetCursorPos(&pt);
+
+                SetForegroundWindow(hWnd); // Required for menu behavior
+
+
+                if (const HMENU hMenu = CreatePopupMenu()) {
+                    const wstring v = L"MusicPP V" + kVersion;
+                    AppendMenu(hMenu, MF_STRING | MF_DISABLED, 0, v.c_str());
+                    AppendMenu(hMenu, MF_STRING, ID_PURGE_DATABASE, L"Purge Database");
+                    AppendMenu(hMenu, MF_SEPARATOR, 0, nullptr);
+
+                    const Snapshot metadata = ctx->player->GetSnapshot(kSnapshotTypeTray);
+                    const wstring title = metadata.title;
+                    const wstring artist = metadata.artist;
+                    const wstring album = metadata.album;
+                    if (metadata.has_session && !title.empty() && !artist.empty() && !album.empty()) {
+                        AppendMenu(hMenu, MF_STRING, ID_COPY_TITLE_ARTIST_ALBUM, L"Now Playing");
+                        AppendMenu(hMenu, MF_STRING, ID_COPY_TITLE, EscapeAmpersands(Truncate(title)).c_str());
+                        AppendMenu(hMenu, MF_STRING, ID_COPY_ARTIST, EscapeAmpersands(Truncate(artist)).c_str());
+                        AppendMenu(hMenu, MF_STRING, ID_COPY_ALBUM, EscapeAmpersands(Truncate(album)).c_str());
+                        AppendMenu(hMenu, MF_SEPARATOR, 0, nullptr);
+                        AppendMenu(hMenu, MF_STRING, ID_COPY_IMAGE, L"Copy Image");
+                        AppendMenu(hMenu, MF_STRING, ID_OPEN_IMAGE, L"Open Image");
+                        AppendMenu(hMenu, MF_STRING, ID_TRAY_FORCE_REFRESH, L"Force Refresh");
+                        AppendMenu(hMenu, MF_SEPARATOR, 0, nullptr);
+                    }
+
+                    const wstring discord_state =
+                            ctx->discord->GetState() ? L"Discord active" : L"Discord disabled";
+                    AppendMenu(hMenu, MF_STRING, ID_TRAY_DISCORD_TOGGLE, discord_state.c_str());
+
+                    AppendMenu(hMenu, MF_STRING, ID_TRAY_LASTFM_TOGGLE, ctx->lastfm->GetReason().c_str());
+
+                    AppendMenu(hMenu, MF_SEPARATOR, 0, nullptr);
+                    AppendMenu(hMenu, MF_STRING, ID_TRAY_EXIT, TEXT("Exit"));
+                    const UINT cmd = TrackPopupMenu(
+                        hMenu,
+                        TPM_RIGHTBUTTON | TPM_BOTTOMALIGN | TPM_RETURNCMD,
+                        pt.x,
+                        pt.y,
+                        0,
+                        hWnd,
+                        nullptr);
+
+                    DestroyMenu(hMenu);
+
+                    if (cmd != 0) {
+                        PostMessage(hWnd, WM_COMMAND, MAKEWPARAM(cmd, 0), 0);
+                    }
+                }
+
+                PostMessage(hWnd, WM_NULL, 0, 0);
+            }
+
+            return 0;
+        }
+
+
+        case WM_COMMAND: {
+            switch (LOWORD(wParam)) {
+                case ID_PURGE_DATABASE: {
+                    if (ctx->logger) ctx->logger->info("User requested database purge");
+                    setup::PurgeDatabase(ctx);
+                    return 0;
+                }
+
+                case ID_TRAY_EXIT: {
+                    if (ctx->logger) ctx->logger->info("User requested application exit via tray menu.");
+                    SendMessage(hWnd, WM_CLOSE, 0, 0);
+                    return 0;
+                }
+
+                case ID_TRAY_LASTFM_TOGGLE: {
+                    if (ctx->logger) ctx->logger->info("LastFM state toggled");
+                    ctx->lastfm->toggle();
+                    return 0;
+                }
+
+                case ID_TRAY_DISCORD_TOGGLE: {
+                    if (ctx->logger) ctx->logger->info("Discord state toggled");
+                    ctx->discord->toggle();
+                    return 0;
+                }
+
+                case ID_COPY_TITLE: {
+                    const Snapshot metadata = ctx->player->GetSnapshot(kSnapshotTypeTray);
+                    CopyToClipboard(ctx, metadata.title, L"Copied title: " + metadata.title);
+                    return 0;
+                }
+
+                case ID_COPY_ARTIST: {
+                    const Snapshot metadata = ctx->player->GetSnapshot(kSnapshotTypeTray);
+                    CopyToClipboard(ctx, metadata.artist, L"Copied artist: " + metadata.artist);
+                    return 0;
+                }
+
+                case ID_COPY_ALBUM: {
+                    const Snapshot metadata = ctx->player->GetSnapshot(kSnapshotTypeTray);
+                    CopyToClipboard(ctx, metadata.album, L"Copied album: " + metadata.album);
+                    return 0;
+                }
+
+                case ID_COPY_IMAGE: {
+                    const Snapshot metadata = ctx->player->GetSnapshot(kSnapshotTypeTray);
+                    CopyToClipboard(ctx, metadata.image,
+                                    L"Copied image URL\nSourced from: " + metadata.image_source);
+                    return 0;
+                }
+
+                case ID_OPEN_IMAGE: {
+                    const Snapshot metadata = ctx->player->GetSnapshot(kSnapshotTypeTray);
+                    ShellExecuteW(nullptr, L"open", metadata.image.c_str(), nullptr, nullptr, SW_SHOWNORMAL);
+                    return 0;
+                }
+
+                case ID_TRAY_FORCE_REFRESH: {
+                    ctx->player->ImageRefresh();
+                    return 0;
+                }
+
+                case ID_COPY_TITLE_ARTIST_ALBUM: {
+                    const Snapshot metadata = ctx->player->GetSnapshot(kSnapshotTypeTray);
+                    const wstring clipboard_text =
+                            metadata.title + L"\r\n" + metadata.artist + L"\r\n" + metadata.album;
+                    const wstring label = L"Copied metadata\nTitle: " + metadata.title + L"\nArtist: " + metadata.
+                                          artist + L"\nAlbum: " + metadata.album;
+                    CopyToClipboard(ctx, clipboard_text, label);
+                    return 0;
+                }
+
+                default: {
+                    return 0;
+                }
+            }
+        }
+
+        case WM_POWERBROADCAST:
+            if (wParam == PBT_APMRESUMEAUTOMATIC || wParam == PBT_APMRESUMESUSPEND) {
+                ctx->spotify->NotifyAwake();
+            }
+
+        case WM_TIMER: {
+            if (!ctx || !ctx->player || !ctx->discord) {
+                if (ctx && ctx->logger) ctx->logger->error("Timer triggered but AppContext/dependencies are null.");
+                return 0;
+            }
+
+            ctx->player->UpdateInfo();
+            ctx->discord->update();
+            UpdateTrayTooltip(ctx);
+            if (ctx->logger) ctx->logger->debug("Data timer update cycle completed.");
+            return 0;
+        }
+
+
+        case WM_CLOSE:
+
+            if (ctx) {
+                KillTimer(hWnd, DATA_TIMER_ID);
+
+                Shell_NotifyIcon(NIM_DELETE, &ctx->nid);
+
+                if (ctx->nid.hIcon) {
+                    DestroyIcon(ctx->nid.hIcon);
+                }
+                if (ctx->logger) ctx->logger->info("Timers and icon cleaned up");
+            }
+
+            DestroyWindow(hWnd);
+
+            return 0;
+
+
+        case WM_DESTROY:
+
+            PostQuitMessage(0);
+
+            return 0;
+
+        default: ;
+    }
+
+    return DefWindowProc(hWnd, msg, wParam, lParam);
+}
+
+int loop::loop(AppContext &ctx, const HINSTANCE hInstance) {
+    g_wm_taskbar_restart = RegisterWindowMessage(TEXT("TaskbarCreated"));
+    if (g_wm_taskbar_restart == 0) {
+        if (ctx.logger) ctx.logger->error("Failed to register TaskbarCreated message");
+    }
+
+    WNDCLASSW wc{};
+    wc.lpfnWndProc = WndProc;
+    wc.hInstance = hInstance;
+    wc.lpszClassName = L"MusicPPTrayClass";
+    wc.hCursor = LoadCursor(nullptr, IDC_ARROW);
+    if (!RegisterClassW(&wc)) return 0;
+    ctx.hWnd = CreateWindowExW(0, L"MusicPPTrayClass", L"TrayOnlyWindow",
+                               WS_OVERLAPPEDWINDOW, CW_USEDEFAULT, CW_USEDEFAULT,
+                               300, 200, nullptr, nullptr, hInstance, &ctx);
+    if (!ctx.hWnd) return 0;
+    ShowWindow(ctx.hWnd, SW_HIDE);
+
+    MSG msg{};
+    while (GetMessage(&msg, nullptr, 0, 0)) {
+        TranslateMessage(&msg);
+        DispatchMessage(&msg);
+    }
+    if (ctx.logger) ctx.logger->info("Application message loop ended. Shutting down.");
+
+    return static_cast<int>(msg.wParam);
+}
